✨ (VoucherController.php, JsonResponseService.php): menambahkan JsonResponseService untuk mengelola respons JSON

+++ database/migrations/2025_02_19_014915_create_vouchers_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('vouchers', function (Blueprint $table) {
             $table->string('id')->primary();
             $table->string('code')->unique();
-            $table->decimal('discount', 5, 2);
+            $table->integer('discount');
             $table->dateTime('expired_at');
             $table->timestamps();
         });
